+++ app/Http/Controllers/DemoproductController.php
@@ -9,32 +9,9 @@ use App\ProductStatus;
 use Illuminate\Http\Request;
 use App\Http\Requests\StoreDemoproduct;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Str;
 
 class DemoproductController extends Controller
 {
-    /**
-     * Recursively populate array with all children
-     * 
-     * @param \App\Location $location
-     * @param mixed $path
-     * @param mixed &$locationBreadcrumbList
-     * @return void 
-     */
-    private function findChildren(Location $location, $path, &$locationBreadcrumbList)
-    {
-        $id = $location->id;
-        foreach ($location->childrenLocations as $children) {
-            $oldPath = $path;
-            $oldId = $id;
-            $path .= ' > ' . $children->name;
-            $this->findChildren($children, $path, $locationBreadcrumbList);
-            $path = $oldPath;
-            $id = $oldId;
-        }
-        $locationBreadcrumbList->push(['name' => $path, 'id' => $id]);
-    }
-
     /**
      * Display a listing of the resource.
      *
@@ -43,13 +20,14 @@ class DemoproductController extends Controller
     public function index()
     {
         $demoproducts = Demoproduct::with('product', 'location', 'status')->get();
-        $selectedproducts = $demoproducts->map(function ( $product ) {
+        $selectedproducts = $demoproducts->map(function ($product) {
             return [
                 'Artikel' => $product->product->item,
                 'Beskrivning' => $product->product->item_description_swe,
                 'Status' => $product->status->description,
                 'Kommentar' => $product->comment,
                 'Plats' => $product->location->path(),
+                'Plats_id' => $product->location_id,
                 'Testad' => $product->tested,
                 'E_nummer' => $product->product->enummer,
                 'Orginal_kartong' => $product->original_box,
@@ -63,14 +41,26 @@ class DemoproductController extends Controller
         });
         $fields = collect([]);
 
-        $fields->push(['key'=> 'Info']);
-        $fields->push(['key'=> 'Artikel']);
-        $fields->push(['key'=> 'Status']);
-        $fields->push(['key'=> 'Plats']);
-        // $filter = request('filter') ;
-        // return view('demoproducts.index', ['products' => $selectedproducts, 'fields' => $fields, 'filter' => $filter]);
-        return view('demoproducts.index', ['products' => $selectedproducts, 'fields' => $fields]);
+        $fields->push(['key' => 'Info']);
+        $fields->push(['key' => 'Artikel']);
+        $fields->push(['key' => 'Status']);
+        $fields->push(['key' => 'Plats']);
+        $locationBreadcrumbList = $this->locationNames();
+        // dd($locationBreadcrumbList);
+        // $locations = Location::all();
+        // $locationBreadcrumbList = $locations->map(function ($f) {
+        //     return [
+        //         'text' => $f->path(),
+        //         'value' => $f->id,
+        //     ];
+        // })->sortBy('text');
+        return view('demoproducts.index', [
+            'products' => $selectedproducts,
+            'fields' => $fields,
+            'locations' => $locationBreadcrumbList,
+        ]);
     }
+
 
     /**
      * Show the form for creating a new resource.
@@ -79,16 +69,9 @@ class DemoproductController extends Controller
      */
     public function create()
     {
-        $locations = Location::whereNull('location_id')
-        ->get();
-        $locationBreadcrumbList = collect([]);
-        foreach ($locations as $parent) {
-            $place = $parent->name;
-            $this->findChildren($parent, $place, $locationBreadcrumbList);
-        }
-        $locationBreadcrumbList = collect($locationBreadcrumbList)->sortBy('name');
+        $locationBreadcrumbList = $this->locationNames();
         $products = Product::limit(50)->get();
-        $items = $products->map(function ( $data ) {
+        $items = $products->map(function ($data) {
             return [
                 'id' => $data->id,
                 'item' => $data->item,
@@ -96,11 +79,11 @@ class DemoproductController extends Controller
             ];
         });
         return view('demoproducts.create', [
-            'locations' => $locationBreadcrumbList, 
+            'locations' => $locationBreadcrumbList,
             'products' => Product::all(),
             'statuses' => ProductStatus::all(),
             'items' => $items,
-            ]);
+        ]);
     }
 
     /**
@@ -118,47 +101,41 @@ class DemoproductController extends Controller
     }
 
     /**
-     * Display the specified resource.
-     *
-     * @param  \App\Demoproduct  $demoproduct
-     * @return \Illuminate\Http\Response
+     * Recursively populate array with all children
+     * 
+     * @param \App\Location $location
+     * @param mixed $path
+     * @param mixed &$names
+     * @return void 
      */
-    public function show(Demoproduct $demoproduct)
+    private function findChildren(Location $location, $path, &$names)
     {
-        //
+        $id = $location->id;
+        foreach ($location->childrenLocations as $children) {
+            $oldPath = $path;
+            $oldId = $id;
+            $path .= ' > ' . $children->name;
+            $this->findChildren($children, $path, $names);
+            $path = $oldPath;
+            $id = $oldId;
+        }
+        $names->push(['name' => $path, 'id' => $id]);
     }
 
     /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Demoproduct  $demoproduct
-     * @return \Illuminate\Http\Response
+     * Create a list as breadcrums
+     * 
+     * @return collection
      */
-    public function edit(Demoproduct $demoproduct)
+    private function locationNames()
     {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Demoproduct  $demoproduct
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, Demoproduct $demoproduct)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Demoproduct  $demoproduct
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy(Demoproduct $demoproduct)
-    {
-        //
+        $locations = Location::whereNull('location_id')
+            ->get();
+        $names = collect([]);
+        foreach ($locations as $parent) {
+            $place = $parent->name;
+            $this->findChildren($parent, $place, $names);
+        }
+        return $names->sortBy('name');
     }
 }