+++ database/factories/TrackDemoproductFactory.php
@@ -1,0 +1,12 @@
+<?php
+
+/** @var \Illuminate\Database\Eloquent\Factory $factory */
+
+use App\TrackDemoproduct;
+use Faker\Generator as Faker;
+
+$factory->define(TrackDemoproduct::class, function (Faker $faker) {
+    return [
+        //
+    ];
+});